AFIRGEN FINAL/tests/requirements-test.txt
# Test dependencies for AFIRGen

# Core testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
hypothesis>=6.92.0  # Property-based testing

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Image processing for test data
Pillow>=10.0.0

# Async support
aiofiles>=23.1.0

# Data validation
pydantic>=2.0.0

# Numpy for embeddings
numpy>=1.24.0

# Database clients
asyncpg>=0.28.0  # For Aurora pgvector
opensearch-py>=2.3.0  # For OpenSearch

# FastAPI for UploadFile
fastapi>=0.100.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
